requirements.txt
pandas
transformers
torch
scikit-learn
prophet
spacy==3.8.0
newsapi-python
matplotlib
seaborn
streamlit
# Direct link to the correct spaCy model wheel file
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl